Expand coldefs to include list of sources for compare table wiget
Because the compare table widget shows results for runs of different
models, it has the tricky problem of dealing with potentially
different sources for any given cell.

The approach here is to provide a complete list of possible sources
for a coldef, which is based on the models defined for the
project. The compare table widget submits a request for all possible
sources and the server returns all the values it can for all of the
runs. The compare table uses the first available value matching a
coldef source in the list of possible sources.

This approach assumes there will never be more than one value for a
given run coldef.