Address suspected race condition
It's been observed that some training operations don't have the last N
events from the TF logs. This is *likely* (unconfirmed) due to a race
at the end of the operation between the file system sync and the last
TF collection. This change calls `sync` before reading from the TF
event collectors.