Manual top-level restart scheme for Guild View
This is rather non-standard for Erlang and not the way we'd typically
run a system, but this is a command line based application and we're
in somewhat uncharted waters. Arguably system process restarts should
be more granular and at the app supervisor level. This tail call to
restart the HTTP server from scratch is the right course given our
current approach.

True servers like Guild Serve for forthcoming Depot should not be
deployed using this pattern.