Hack to avoid loosing data polling for temporary server errors
Eating errors at the lowest level and signalling using null -
terrible. This is a narrowly defined piece of functionality however
and used only by the data listener. It's either this or move the
response handling logic into the data listener, which really just
wants parsed JSON.

We really want parametric dispatch on an error type, which is
basically what we're doing with null.