Seemed-to-have-fixed chart init issues
This appears to have fixed the chart init issue when charts are
displayed on the Summary page when the Summary page is not the initial
page.

The vz-line-chart component wants a particular state when
initializing. TensorBoard charts are initialized after the page is
visible and the user expands a tag group - that works. Summary page
charts display when the Summary page is initially rendered - that
works. When the Summary page is *not* initialy rendered, we get:

> zoomed out too far: could not find suitable interval to display
> labels

from plottable.js:751.

The approach is:

- Wait until guild-chart is 'ready', caching all chart operations
  applied

- When 'ready' apply cached chart operations - but apply them *after*
  the ready event using setTimeout

I think we want a `on-container-visible` event to finally initialize
the chart. Hopefully what we have is the correct proxy.