Take guild out of docs build process
We use Guild to generate a JSON file containing its command info,
which is used in turn to generate docs. The requirement to build Guild
in order to generate docs is a bit much, so we'll store the latest
JSON as source and manually update using the `refresh-commands-json`
make target as needed.