Wip toward operation refactor
The goal is to decouple the operations so we can modify them more
easily and add new ops without thinking too hard about it impacting
the others.

The current implementation is a bit of a mess, as it reflects the
complexities of the runtime abstract, which is no longer in play. We
can now just build our TensorFlow specific ops directly without having
to work through an abstraction layer (guild_operation).