Quick fix for latest TensorBoard support
Note that TensorBoard support in this lineage of Guild (Erlang based) is
quite broken as it relies on a data API that is no longer supported
in more recent versions of TensorBoard. The guild-python (Python based)
repo contains an in-progress work that fully embeds TensorBoard in the
CLI (for Guild View) and eliminates this conflict.

Please note that Guild View in guild-python as of this commit is *not*
fully functional, though this is our current work effort, which we hope
to have working in the next week or so.